requirements.txt
openai-agents
pydantic~=2.11.1
duckduckgo-search
pypdf
beautifulsoup4
arize-phoenix-otel
openai~=1.70.0
gradio
requests
markdownify
openinference-instrumentation-openai